refactor: 워크플로우 리스트 hook 책임 분리
