+++ src/pages/workflows/model/index.ts
@@ -1,4 +1,6 @@
 export * from "./constants";
 export * from "./types";
-export * from "./useWorkflowListSection";
+export * from "./useWorkflowListActions";
+export * from "./useWorkflowListData";
+export * from "./useWorkflowListInfiniteScroll";
 export * from "./workflow-list";
